# public/openapi.yaml
openapi: 3.1.0
info:
  title: Kaspi Automation API
  version: 1.0.0
  description: Tenant-isolated API for Kaspi QR payments, phone invoices, recurring schedules, cancellation, refunds, and safe sandbox simulation.
servers:
  - url: /api/v1
security:
  - bearerApiKey: []
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      summary: List recurring payment schedules
      responses:
        '200':
          description: Recurring schedules
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Subscription' } }
    post:
      operationId: createSubscription
      summary: Create a recurring phone-invoice schedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSubscription' }
      responses:
        '201':
          description: Recurring schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /subscriptions/{subscriptionId}:
    get:
      operationId: getSubscription
      summary: Get a recurring schedule
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: Recurring schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Subscription' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateSubscription
      summary: Update a recurring schedule
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSubscription' }
      responses:
        '200':
          description: Updated recurring schedule
...
  /subscriptions/{subscriptionId}/invoices:
    get:
      operationId: listSubscriptionInvoices
      summary: List generated invoice cycles
      parameters: [{ $ref: '#/components/parameters/SubscriptionId' }]
      responses: { '200': { description: Invoice cycles } }
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: kpa_live
      description: API key created in the customer dashboard.
  parameters:
...
    SubscriptionId:
      name: subscriptionId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    CreateSubscription:
      type: object
      required: [customerPhone, amount, billingPeriod, firstPaymentAt]
      properties:
        customerName: { type: string, maxLength: 200 }
        customerPhone: { type: string, pattern: '^7\d{10}$' }
        amount: { type: number, exclusiveMinimum: 0 }
        billingPeriod: { type: string, enum: [daily, weekly, biweekly, monthly, quarterly, yearly] }
        billingDay: { type: integer, minimum: 1, maximum: 28 }
        billingTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
        firstPaymentAt: { type: string, format: date-time }
        totalCycles: { type: [integer, 'null'], minimum: 1 }
        maxRetryAttempts: { type: integer, minimum: 0, maximum: 10, default: 3 }
        retryIntervalHours: { type: integer, minimum: 1, maximum: 168, default: 24 }
        gracePeriodDays: { type: integer, minimum: 0, maximum: 30, default: 3 }
        description: { type: string, maxLength: 60 }
    Subscription:
      type: object
      required: [id, customerPhone, amount, currency, billingPeriod, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        customerName: { type: [string, 'null'] }
        customerPhone: { type: string }
        amount: { type: number }
        currency: { type: string, const: KZT }
        billingPeriod: { type: string }
        billingDay: { type: [integer, 'null'] }
        billingTime: { type: string }
        status: { type: string, enum: [active, paused, completed, cancelled] }
        nextPaymentAt: { type: [string, 'null'], format: date-time }
        totalCycles: { type: [integer, 'null'] }
        generatedCycles: { type: integer }
        successfulCycles: { type: integer }
        createdAt: { type: string, format: date-time }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        requestId: { type: string }
  responses:
...
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotFound:
      description: Resource not found in this tenant
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...